Refactor admin routes into modules
